pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cell-gnn"
version = "0.1.0"
description = "Graph Neural Networks for cell simulations (arbitrary, boids, gravity)"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy",
    "scipy",
    "matplotlib",
    "scikit-learn",
    "pydantic>=2.0",
    "pyyaml",
    "tqdm",
    "imageio",
    "umap-learn",
    "geomloss",
    "prettytable",
    "tifffile",
    "seaborn",
]

[tool.setuptools.packages.find]
where = ["src"]
